+++ bofhound/ad/models/__init__.py
@@ -7,3 +7,4 @@ from .bloodhound_schema import BloodHoundSchema
 from .bloodhound_ou import BloodHoundOU
 from .bloodhound_gpo import BloodHoundGPO
 from .bloodhound_domaintrust import BloodHoundDomainTrust
+from .bloodhound_crossref import BloodHoundCrossRef